refactor: added onException joinPoint as a feature on top vanilla kaop
