+++ package.json
@@ -92,6 +92,6 @@
     "validate-commit-msg": "^2.10.1"
   },
   "dependencies": {
-    "kaop": "^3.0.1"
+    "kaop": "^3.1.0"
   }
 }
