+++ src/index.ts
@@ -1,2 +1,3 @@
 export { afterMethod, beforeMethod, afterInstance, beforeInstance } from "./decorators"
 export { AdviceRef, ClassSignature, Metadata, MethodSignature } from "./interfaces"
+export { onException } from "./onException"
